Current pid is alive

+++ lib/scout/misc/monitor.rb
@@ -1,5 +1,6 @@
 module Misc
   def self.pid_alive?(pid)
+    return true if Process.pid == pid
     !! Process.kill(0, pid) rescue false
   end
